+++ test/unit/fuzz-tools.test.js
@@ -259,6 +259,7 @@ import * as inflamV268 from '../../lib/inflam-v268.js';
 import * as metabolicV269 from '../../lib/metabolic-v269.js';
 import * as adiposityV270 from '../../lib/adiposity-v270.js';
 import * as lipidsV271 from '../../lib/lipids-v271.js';
+import * as anthroV272 from '../../lib/anthro-v272.js';
 
 const MODULES = {
   'clinical.js': clinical,
@@ -462,6 +463,7 @@ const MODULES = {
   'metabolic-v269.js': metabolicV269,
   'adiposity-v270.js': adiposityV270,
   'lipids-v271.js': lipidsV271,
+  'anthro-v272.js': anthroV272,
 };
 
 const MATRIX = [0, -1, 1e9, NaN, Infinity, -Infinity, '', undefined, null];